+++ lib/form_forms.rb
@@ -2,7 +2,7 @@ require 'form_forms/mixins'
 require 'form_forms/elements'
 require 'form_forms/forms'
 
-require 'form_forms/form_registry'
+require 'form_forms/registry'
 
 require "form_forms/version"