+++ lib/form_forms/registry.rb
@@ -1,7 +1,7 @@
 require 'active_support/hash_with_indifferent_access'
 
 module FormForms
-  class FormRegistry
+  class Registry
     class << self
       def method_missing(method, *args, &block)
         if forms.respond_to?(method)
@@ -14,6 +14,16 @@ module FormForms
     protected
       def forms
         @forms ||= HashWithIndifferentAccess.new
+      end
+    end
+  end
+
+  # Deprecated class
+  class FormRegistry
+    class << self
+      def method_missing(method, *args, &block)
+        ::ActiveSupport::Deprecation.warn("FormRegistry is deprecated and will be removed from FormForms in version 0.4.0", caller)
+        Registry.send(method, *args, &block)
       end
     end
   end