+++ test/form_forms/forms/form_test.rb
@@ -4,11 +4,11 @@ module FormForms
   module Forms
     class FormTest < ActionView::TestCase
       test "register form" do
-        assert_nil FormForms::FormRegistry['form_test']
+        assert_nil FormForms::Registry['form_test']
 
         form = Form.register('form_test'){}
 
-        assert_equal FormForms::FormRegistry['form_test'], form
+        assert_equal FormForms::Registry['form_test'], form
       end
     end
   end